# Bind `stdouttrace` self-observability instruments