… exporters/stdout/stdouttrace/go.sum