Alternative to https://github.com/open-telemetry/opentelemetry-go/pull/7215

Base on changes in https://github.com/open-telemetry/opentelemetry-go/pull/7222 and https://github.com/open-telemetry/opentelemetry-go/pull/7223.

This uses the [github.com/MrAlias/bind](https://github.com/MrAlias/bind) package to bind reused attributes during instrument creation. This efficiently handles sending these values on the hot-path.

Additionally, the combination of #7223 and the `bind` package allow the error case to be as optimized as possible. The remaining allocations are from the new set allocations like in #7215.  The difference here compared to #7215 is no additional pools are needed and the call-sites are not verbose.

### Benchmarks

```console
$ benchstat main_d99c68cb2.txt poc-bind-telemetry-stdouttrace_bfbff08b1.txt
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/exporters/stdout/stdouttrace
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
                                        │ main_d99c68cb2.txt │ poc-bind-telemetry-stdouttrace_bfbff08b1.txt │
                                        │       sec/op       │        sec/op          vs base               │
ExporterExportSpans/SelfObservability-8          27.12µ ± 3%             26.90µ ± 3%       ~ (p=0.481 n=10)
ExporterExportSpans/NoObservability-8            26.96µ ± 2%             26.71µ ± 2%       ~ (p=0.190 n=10)
geomean                                          27.04µ                  26.80µ       -0.86%

                                        │ main_d99c68cb2.txt │ poc-bind-telemetry-stdouttrace_bfbff08b1.txt │
                                        │        B/op        │        B/op          vs base                 │
ExporterExportSpans/SelfObservability-8         4.081Ki ± 0%          3.974Ki ± 0%  -2.63% (p=0.000 n=10)
ExporterExportSpans/NoObservability-8           3.873Ki ± 0%          3.873Ki ± 0%       ~ (p=1.000 n=10) ¹
geomean                                         3.976Ki               3.923Ki       -1.32%
¹ all samples are equal

                                        │ main_d99c68cb2.txt │ poc-bind-telemetry-stdouttrace_bfbff08b1.txt │
                                        │     allocs/op      │      allocs/op       vs base                 │
ExporterExportSpans/SelfObservability-8           67.00 ± 0%            68.00 ± 0%  +1.49% (p=0.000 n=10)
ExporterExportSpans/NoObservability-8             65.00 ± 0%            65.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                           65.99                 66.48       +0.74%
¹ all samples are equal
```