@flc1125 2025-08-27
If we refer to the `tracer`, we could pre-set some values at creation time, like this:

```go
package main

import (
	"context"

	"go.opentelemetry.io/otel/attribute"
	"go.opentelemetry.io/otel/metric"
	"go.opentelemetry.io/otel/trace"
)

func main() {
	var (
		tracer trace.Tracer
		meter metric.Meter
	)

	tracer.Start(
		context.TODO(),
		"operationName",
		trace.WithAttributes(attribute.String("key", "value")), // <--- Refer to a design similar to this one
	)

	meter.Float64Counter(
		"example.com/metric",
		metric.WithDescription("An example metric"),
		metric.WithAttributeSet(attribute.NewSet(...)), //
		metric.WithAttributes(attribute.String()...)),
	)
}
```