@MrAlias 2025-08-27
> If we refer to the `tracer`, we could pre-set some values at creation time, like this:
> 
> ```go
> package main
> 
> import (
> 	"context"
> 
> 	"go.opentelemetry.io/otel/attribute"
> 	"go.opentelemetry.io/otel/metric"
> 	"go.opentelemetry.io/otel/trace"
> )
> 
> func main() {
> 	var (
> 		tracer trace.Tracer
> 		meter metric.Meter
> 	)
> 
> 	tracer.Start(
> 		context.TODO(),
> 		"operationName",
> 		trace.WithAttributes(attribute.String("key", "value")), // <--- Refer to a design similar to this one
> 	)
> 
> 	meter.Float64Counter(
> 		"example.com/metric",
> 		metric.WithDescription("An example metric"),
> 		metric.WithAttributeSet(attribute.NewSet(...)), //
> 		metric.WithAttributes(attribute.String()...)),
> 	)
> }
> ```

Attributes are not a part of an instrument's definition: https://github.com/open-telemetry/opentelemetry-go/blob/metric/v1.37.0/metric/syncfloat64.go#L32

We would first need to resolution on https://github.com/open-telemetry/opentelemetry-specification/issues/4126 before we could change the API.