@MrAlias 2025-08-29
Closing as the error case requires an additional set allocation.